# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sea Gull Restaurant API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface RegisterRequest {
  name: string;
  phone: string;
  email?: string;
  password?: string;
}

export interface LoginRequest {
  phone: string;
  password?: string;
}

export interface User {
  id: number;
  name: string;
  phone: string;
  email?: string;
  role: string;
  avatarUrl?: string;
  loyaltyPoints: number;
  preferredLanguage?: string;
  darkMode?: boolean;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface Category {
  id: number;
  name: string;
  nameAr?: string;
  description?: string;
  imageUrl?: string;
  icon?: string;
  sortOrder?: number;
  isActive?: boolean;
}

export interface Product {
  id: number;
  categoryId: number;
  name: string;
  nameAr?: string;
  description?: string;
  price: string;
  discountedPrice?: string;
  imageUrl?: string;
  images?: string[];
  prepTime?: number;
  spiceLevel?: string;
  isAvailable?: boolean;
  isFeatured?: boolean;
  isBestSeller?: boolean;
  isNew?: boolean;
  isVegetarian?: boolean;
  isHealthy?: boolean;
  tags?: string[];
  rating?: string;
  reviewCount?: number;
  calories?: number;
}

export interface AddOn {
  id: number;
  name: string;
  nameAr?: string;
  price: string;
  isAvailable?: boolean;
}

export interface ProductVariant {
  id: number;
  name: string;
  nameAr?: string;
  priceDiff?: string;
  isDefault?: boolean;
}

export type ProductDetail = Product & {
  ingredients?: string;
  allergens?: string;
  addOns?: AddOn[];
  variants?: ProductVariant[];
  relatedProducts?: Product[];
};

export type OrderItemAddOnsItem = { [key: string]: unknown };

export interface OrderItem {
  id: number;
  productId?: number;
  productName: string;
  productImage?: string;
  variantName?: string;
  quantity: number;
  unitPrice: string;
  totalPrice: string;
  addOns?: OrderItemAddOnsItem[];
  specialInstructions?: string;
}

export interface StatusHistory {
  status: string;
  note?: string;
  createdAt: string;
}

export interface Branch {
  id: number;
  name: string;
  nameAr?: string;
  address: string;
  area?: string;
  city?: string;
  phone?: string;
  latitude?: string;
  longitude?: string;
  isActive?: boolean;
  acceptsDelivery?: boolean;
  acceptsPickup?: boolean;
}

export interface RiderInfo {
  id?: number;
  name?: string;
  phone?: string;
  rating?: string;
  vehicleType?: string;
  profileImageUrl?: string;
  currentLatitude?: string;
  currentLongitude?: string;
}

export interface Order {
  id: number;
  orderNumber: string;
  status: string;
  type?: string;
  subtotal?: string;
  deliveryFee?: string;
  serviceFee?: string;
  tax?: string;
  tip?: string;
  discount?: string;
  total: string;
  deliveryAddress?: string;
  deliveryNotes?: string;
  specialInstructions?: string;
  couponCode?: string;
  loyaltyPointsEarned?: number;
  loyaltyPointsUsed?: number;
  estimatedDeliveryTime?: string;
  createdAt: string;
  items?: OrderItem[];
  statusHistory?: StatusHistory[];
  branch?: Branch;
  rider?: RiderInfo;
}

export type CreateOrderRequestItemsItem = { [key: string]: unknown };

export interface CreateOrderRequest {
  branchId: number;
  items: CreateOrderRequestItemsItem[];
  paymentMethod: string;
  type: string;
  addressId?: number;
  deliveryAddress?: string;
  deliveryNotes?: string;
  specialInstructions?: string;
  couponCode?: string;
  loyaltyPointsToUse?: number;
  tip?: number;
  subtotal: number;
  deliveryFee?: number;
  serviceFee?: number;
  tax?: number;
  total: number;
  scheduledFor?: string;
  isContactless?: boolean;
}

export interface Address {
  id: number;
  label: string;
  address: string;
  area?: string;
  city?: string;
  latitude?: string;
  longitude?: string;
  landmark?: string;
  floorApartment?: string;
  isDefault?: boolean;
}

export interface AddressRequest {
  label: string;
  address: string;
  area?: string;
  city?: string;
  latitude?: string;
  longitude?: string;
  landmark?: string;
  floorApartment?: string;
  isDefault?: boolean;
}

export interface Notification {
  id: number;
  type: string;
  title: string;
  body: string;
  orderId?: number;
  isRead?: boolean;
  createdAt: string;
}

export interface CouponValidation {
  valid: boolean;
  discount?: number;
  message?: string;
  couponId?: number;
}

export type GetProductsParams = {
  categoryId?: number;
  search?: string;
  featured?: boolean;
  bestSeller?: boolean;
  vegetarian?: boolean;
  healthy?: boolean;
};

export type CancelOrderBody = {
  reason?: string;
};

export type ValidateCouponBody = {
  code: string;
  orderAmount: number;
};