# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Sea Gull Restaurant API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    get:
      operationId: getOrders
      tags: [orders]
      summary: Get user orders
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
...
components:
  schemas:
...
    Order:
      type: object
      required: [id, orderNumber, status, total, createdAt]
      properties:
        id:
          type: number
        orderNumber:
          type: string
        status:
          type: string
        type:
          type: string
        subtotal:
          type: string
        deliveryFee:
          type: string
        serviceFee:
          type: string
        tax:
          type: string
        tip:
          type: string
        discount:
          type: string
        total:
          type: string
        deliveryAddress:
          type: string
        deliveryNotes:
          type: string
        specialInstructions:
          type: string
        couponCode:
          type: string
        loyaltyPointsEarned:
          type: number
        loyaltyPointsUsed:
          type: number
        estimatedDeliveryTime:
          type: string
        createdAt:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        statusHistory:
          type: array
          items:
            $ref: "#/components/schemas/StatusHistory"
        branch:
          $ref: "#/components/schemas/Branch"
        rider:
          $ref: "#/components/schemas/RiderInfo"
    OrderItem:
      type: object
      required: [id, productName, quantity, unitPrice, totalPrice]
      properties:
        id:
          type: number
        productId:
          type: number
        productName:
          type: string
        productImage:
          type: string
        variantName:
          type: string
        quantity:
          type: number
        unitPrice:
          type: string
        totalPrice:
          type: string
        addOns:
          type: array
          items:
            type: object
        specialInstructions:
          type: string
    StatusHistory:
      type: object
      required: [status, createdAt]
      properties:
        status:
          type: string
        note:
          type: string
        createdAt:
          type: string
    RiderInfo:
      type: object
      properties:
        id:
          type: number
        name:
          type: string
        phone:
          type: string
        rating:
          type: string
        vehicleType:
          type: string
        profileImageUrl:
          type: string
        currentLatitude:
          type: string
        currentLongitude:
          type: string
    CreateOrderRequest:
      type: object
      required: [branchId, items, paymentMethod, type, subtotal, total]
      properties:
        branchId:
          type: number
        items:
          type: array
          items:
            type: object
        paymentMethod:
          type: string
        type:
          type: string
        addressId:
          type: number
        deliveryAddress:
          type: string
        deliveryNotes:
          type: string
        specialInstructions:
          type: string
        couponCode:
          type: string
        loyaltyPointsToUse:
          type: number
        tip:
          type: number
        subtotal:
          type: number
        deliveryFee:
          type: number
        serviceFee:
          type: number
        tax:
          type: number
        total:
          type: number
        scheduledFor:
          type: string
        isContactless:
          type: boolean
...
    Branch:
      type: object
      required: [id, name, address]
      properties:
        id:
          type: number
        name:
          type: string
        nameAr:
          type: string
        address:
          type: string
        area:
          type: string
        city:
          type: string
        phone:
          type: string
        latitude:
          type: string
        longitude:
          type: string
        isActive:
          type: boolean
        acceptsDelivery:
          type: boolean
        acceptsPickup:
          type: boolean
...